mini_project/requirements.txt
python-dotenv
yt-dlp
pytube
moviepy
Pillow
google-generativeai
aiofiles
youtube_transcript_api